requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pymongo==4.6.1
python-multipart==0.0.6
face-recognition==1.3.0
numpy==1.24.4
opencv-python-headless==4.8.1.78
Pillow==10.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
websockets==12.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
reportlab==4.1.0
openpyxl==3.1.2
email-validator==2.1.0
